binBoundaries ArrayList<Number> ivm state uit GWT/HTML5, ihkv DWOWIDGET-65
git-svn-id: https://svn.science.uu.nl/repos/project.fisme.java/Statistiek/trunk@15445 58dd2db4-0f11-0410-9df4-da15f00f50bd

+++ src/fi/statistiek/boxplot/BoxplotUserOptionsPanel.java
@@ -606,17 +606,17 @@ public class BoxplotUserOptionsPanel extends JPanel implements ActionListener
 				|| splitType.equals(AllowedTypes.INTEGER))
 			{
 				this.splitMinBoundaryField.setText(
-					Statistiek.getStringValue(this.model.getSplitOptions().getBinBoundaries().get(0)));
+					Statistiek.getStringValue(this.model.getSplitOptions().getBinBoundaries().get(0).doubleValue()));
 				this.splitBinWidthField.setText(Statistiek.getFormattedBinWidth(this.model.getSplitBinBoundaries()));
 				StringBuilder sb = new StringBuilder();
 				for (int i = 0; i < this.model.getSplitOptions()
 					.getBinBoundaries().size() - 1; i++)
 				{
 					sb.append(
-						Statistiek.getStringValue(this.model.getSplitOptions().getBinBoundaries().get(i)));
+						Statistiek.getStringValue(this.model.getSplitOptions().getBinBoundaries().get(i).doubleValue()));
 					sb.append(" -< ");
 					sb.append(
-						Statistiek.getStringValue(this.model.getSplitOptions().getBinBoundaries().get(i + 1)));
+						Statistiek.getStringValue(this.model.getSplitOptions().getBinBoundaries().get(i + 1).doubleValue()));
 					sb.append("\n");
 				}
 				//sb.delete(sb.length() - 3, sb.length());
